test: fix svgo case
