+++ e2e/cases/svg/svgo-minify-view-box/rsbuild.config.ts
@@ -1,0 +1,11 @@
+import { pluginReact } from '@rsbuild/plugin-react';
+import { pluginSvgr } from '@rsbuild/plugin-svgr';
+
+export default {
+  plugins: [
+    pluginReact(),
+    pluginSvgr({
+      svgDefaultExport: 'url',
+    }),
+  ],
+};
